Fix issue with Firefox HTML parsing
The chat message that is emitted, for some reason, is rendered with an
`<img>` tag that does not self-close properly. The Handlebars file has a
properly `/>`-closed tag, but for whatever reason, that tag is rendered
without the self-close. Chrome handles this fine, but Firefox is much
more strict about its XML parsing when using DOMParser.parseFromString.

When parsing the string as `text/html` instead, both browsers work fine.