fix(lloyds): avoid crash on non-unique random points
The algorithm needs unique seed points to run correctly so instead of
generating points which may be duplicates, get a unique list of all
input points and then shuffle that list, and use the first k points.

If this is not a sufficient number of points, extend the list with some
synthetic unique additions by taking the largest X and Y values and
adding an offset.