feat(flow): add effect-guessing heuristic for unknown weapons
If we do not have a defined effect for a weapon, allow the module to
pick one which should be appropriate. This helps to bridge the gap for
homebrew items in the absence of a customization option.

As this may not always be desirable, gate the heuristic behind a config
option (enabled by default).

All credit to @dodgepong for the idea and algorithm.