feat(manager): add typeahead to "Lancer ID" field
Add native browser typeahead to the "Lancer ID" field in the effects
manager. This reduces the need to reference the README when making
custom FX for special LIDs.

Note that we do not scan the user's game for every available LID, as
this could be very slow, and would mean loading every compendium. We
approximate the list by using the keys in our `weaponEffects` map
instead.