feat(build): add bundler, file watcher, and hot-reload
This allows live changes to CSS files and Handlebars templates. Other
files will be automatically copied, but require a refresh of Foundry
to update.

Compendium packs must still be re-built when no world is active, but
this is now documented.

This also allows the module's source files to be moved out of the
`modules/` folder, which avoids the risk of updating all modules and
having local changes be overwritten.